Refactor vote submission into its own class/file
