+++ src/electionguard_cli/e2e_steps/__init__.py
@@ -1,7 +1,9 @@
 from electionguard_cli.e2e_steps.key_ceremony_step import KeyCeremonyStep
 from electionguard_cli.e2e_steps.election_builder_step import ElectionBuilderStep
+from electionguard_cli.e2e_steps.submit_votes_step import SubmitVotesStep
 
 __all__ = [
     "ElectionBuilderStep",
-    "KeyCeremonyStep"
+    "KeyCeremonyStep",
+    "SubmitVotesStep"
 ]
